ファイル整理後の完成版
