+++ docker-compose.yml
@@ -11,6 +11,8 @@ services:
     command: air
     depends_on:
       - db
+    env_file:
+      - .env
 
   frontend:
     build: ./frontend
